initial work for classpath refactor #299
