+++ bundles/com.salesforce.bazel.eclipse.common/src/main/java/com/salesforce/bazel/eclipse/classpath/BazelClasspathContainer.java
@@ -53,9 +53,9 @@ import com.salesforce.bazel.eclipse.component.ComponentContext;
 import com.salesforce.bazel.eclipse.runtime.api.JavaCoreHelper;
 import com.salesforce.bazel.eclipse.runtime.api.ResourceHelper;
 import com.salesforce.bazel.sdk.command.BazelCommandLineToolConfigurationException;
-import com.salesforce.bazel.sdk.lang.jvm.BazelJvmClasspath;
-import com.salesforce.bazel.sdk.lang.jvm.BazelJvmClasspathResponse;
-import com.salesforce.bazel.sdk.lang.jvm.DynamicBazelJvmClasspath;
+import com.salesforce.bazel.sdk.lang.jvm.classpath.BazelJvmClasspath;
+import com.salesforce.bazel.sdk.lang.jvm.classpath.BazelJvmClasspathResponse;
+import com.salesforce.bazel.sdk.lang.jvm.classpath.DynamicBazelJvmClasspath;
 import com.salesforce.bazel.sdk.logging.LogHelper;
 import com.salesforce.bazel.sdk.model.BazelWorkspace;
 import com.salesforce.bazel.sdk.project.BazelProjectManager;
